Crate glib_unix

Rust GLib Unix bindings

Rust bindings and wrappers for glib, part of gtk-rs-core.

GLib 2.56 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.83.0.

Documentation

Rust API - Stable

Rust API - Development

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
glib = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "glib" }

Avoid mixing versioned and git crates like this:

# This will not compile
[dependencies]
glib = "0.13"
glib = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "glib" }

See Also

gio

License

glib-unix is available under the MIT License, please refer to it.

Re-exports

pub use glib_unix_sys as ffi;

Modules

prelude

Functions

closefrom

error_quark

fd_source_new

Adds a closure to be called by the main loop the returned Source is attached to whenever a UNIX file descriptor reaches the given IO condition.

signal_future

Create a Future that will resolve once the given UNIX signal is raised

signal_future_with_priorityfutures

Create a Future that will resolve once the given UNIX signal is raised

signal_source_new

Adds a closure to be called by the main loop the returned Source is attached to whenever a UNIX signal is raised.

signal_streamfutures

Create a Stream that will provide a value whenever the given UNIX signal is raised

signal_stream_with_priorityfutures

Create a Stream that will provide a value whenever the given UNIX signal is raised

unix_fd_add

Adds a closure to be called by the main loop the returned Source is attached to whenever a UNIX file descriptor reaches the given IO condition.

unix_fd_add_full

Adds a closure to be called by the main loop the returned Source is attached to whenever a UNIX file descriptor reaches the given IO condition.

unix_fd_add_local

Adds a closure to be called by the main loop the returned Source is attached to whenever a UNIX file descriptor reaches the given IO condition.

unix_fd_add_local_full

Adds a closure to be called by the main loop the returned Source is attached to whenever a UNIX file descriptor reaches the given IO condition.

unix_open_pipe

unix_signal_add

Adds a closure to be called by the default main loop whenever a UNIX signal is raised.

unix_signal_add_local

Adds a closure to be called by the default main loop whenever a UNIX signal is raised.

unix_signal_add_local_once

Adds a closure to be called by the default main loop whenever a UNIX signal is raised.

unix_signal_add_once

Adds a closure to be called by the default main loop whenever a UNIX signal is raised.